requirements/common.txt
wagtail>=1.8,<1.11
django-bootstrap3>=8.0,<9.0
django-embed-video>=1.1,<1.2
django-fontawesome>=0.2,<0.3
psycopg2>=2.6
